+++ Avalon/Avalon.Draw/GradientSpreadList.cs
@@ -38,7 +38,7 @@ public class GradientSpreadList : Any
 
     protected virtual GradientSpread AddItem(ulong o)
     {
-        int index;
+        long index;
         index = this.Index;
 
         GradientSpread item;
@@ -61,13 +61,13 @@ public class GradientSpreadList : Any
 
     protected virtual Array Array { get; set; }
 
-    protected virtual int ArrayCount { get { return 3; } set { } }
+    protected virtual long ArrayCount { get { return 3; } set { } }
 
-    public virtual int Count { get; set; }
+    public virtual long Count { get; set; }
     
-    protected virtual int Index { get; set; }
+    protected virtual long Index { get; set; }
 
-    public virtual GradientSpread Get(int index)
+    public virtual GradientSpread Get(long index)
     {
         return (GradientSpread)this.Array.GetAt(index);
     }
