+++ Avalon/Avalon.Draw/TextAlign.cs
@@ -2,6 +2,6 @@ namespace Avalon.Draw;
 
 public class TextAlign : Any
 {
-    public virtual int Index { get; set; }
+    public virtual long Index { get; set; }
     internal virtual ulong Intern { get; set; }
 }
